noodlin more

+++ src/toolchain/parser/node.rs
@@ -2,7 +2,7 @@ use crate::toolchain::lexer::tokenized_buffer::TokenizedBuffer;
 
 pub struct Node {
     pub kind: NodeKind,
-    pub token_index: i32,
+    pub token_index: usize,
     pub subtree_size: i32,
 }
